requirements.txt
pytest==7.4.0
requests==2.31.0
structlog~=23.1.0
curlify~=2.2.1
pydantic==2.3.0
pyhamcrest
records~=0.5.3
psycopg2-binary
sqlalchemy~=1.4.49
sqlacodegen
vyper-config
allure-pytest
